Move live transcription to beta
